add in Decimal: 	Decimal operator -(); 	friend Decimal abs(const Decimal&); 	friend Decimal& Power(const Decimal&, double); 	friend bool operator <(const Decimal &a, const Decimal &b); 	friend bool operator <=(const Decimal &a, const Decimal &b); 	friend bool operator ==(const Decimal &a, const Decimal &b); 	Decimal now has the ability to perform reduction of a fraction
add in Integer:
	Integer operator ++();
	Integer operator ++(int);
	friend Integer& IntDivide(const Integer &a, const Integer &b);
	friend Integer& operator %(const Integer &a, const Integer &b);
	friend bool operator <(const Integer &a, const Integer &b);
	friend bool operator ==(const Integer &a, const Integer &b);

+++ Calculator/Calculator.cpp
@@ -848,7 +848,7 @@ void Calculator::InsertCharacter(ButtonCode bc) {
 							{
 								int p = atoi(Value->getNum().c_str());
 								if (PrecValue->getType() == 'd' && Value->getType() == 'i')
-									result = &(Power(*(Integer*)PrecValue, p));
+									result = &(Power(*(Decimal*)PrecValue, p));
 								else if (PrecValue->getType() == 'i' && Value->getType() == 'i')
 									result = &(Power(*(Integer*)PrecValue, p));
 								break;
